% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus-functions.R
\name{CITRUSRegression}
\alias{CITRUSRegression}
\title{CITRUS regression}
\usage{
CITRUSRegression(clustering, outcome)
}
\arguments{
\item{clustering}{A list object returned by RunCITRUS.}

\item{outcome}{A vector of response.}
}
\value{
A regression result object.
}
\description{
Run CITRUS regression function.
}
