% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenograph-functions.R
\name{ConstructAbundanceMatrix}
\alias{ConstructAbundanceMatrix}
\title{Construct Abundance Matrix}
\usage{
ConstructAbundanceMatrix(data)
}
\arguments{
\item{data}{A data frame consisting of three columns: fileid, cluster and
outcome corresponding to each cell.}
}
\description{
Construct abundance matrix given the cluster assignment from PhenoGraph
}
