% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{ExportSubsetFCS}
\alias{ExportSubsetFCS}
\title{Export combinedFCS as FCS files}
\usage{
ExportSubsetFCS(clustering, out.folder, name, desc)
}
\arguments{
\item{clustering}{A list containing combinedFCS object output from citrus.}

\item{out.folder}{A character scalar of full path to output folder.}

\item{name}{A character vector containing selected channel name.}

\item{desc}{A character vector containing corresponding channel description.}
}
\value{
None
}
\description{
Export combinedFCS object from citrus into a FCS file with selected channels.
}
\details{
Exported files will be named as sampled.fcs in the folder specified. The
file that each cells in the exported FCS are indexed and saved as fileID.csv
and corresponding file names of each index are saved as filenames.csv
}
