% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-qPCR.R
\name{ExtractSheetNames}
\alias{ExtractSheetNames}
\title{Extract Excel Sheet Names}
\usage{
ExtractSheetNames(path, pattern = "")
}
\arguments{
\item{path}{A character scalar of full path name of the excel sheet.}

\item{pattern}{An optional regular expression. Only sheet names matching the
pattern will be returned. If not specified, all sheets are returned.}
}
\value{
A character vector of sheet names matching the regular expression
}
\description{
Extract Excel sheet names allowing regular expression for matching.
}
\examples{
\dontrun{FindSheet('path/to/xlsx', '^a-l')}
}
