% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{GetParameters}
\alias{GetParameters}
\title{Get FCS Channel Name and Description}
\usage{
GetParameters(path)
}
\arguments{
\item{path}{A character vector of FCS file full paths.}
}
\value{
A data frame that contains the name and description of the channels
}
\description{
Get the FCS channel name and corresponding description.
}
