% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenograph-functions.R
\name{PhenoGraphTSNE}
\alias{PhenoGraphTSNE}
\title{Run Rtsne on Data Matrix}
\usage{
PhenoGraphTSNE(data, desc.selected = NULL)
}
\arguments{
\item{data}{A data matrix.}

\item{desc.selected}{Names of columns to run Rtsne on.}
}
\value{
A Rtsne object.
}
\description{
Run Rtsne on a data matrix on given columns.
}
