% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso-functions.R
\name{plot.RunLasso}
\alias{plot.RunLasso}
\title{Plot Lasso Coefficient Trace}
\usage{
\method{plot}{RunLasso}(res)
}
\arguments{
\item{res}{A RunLasso object.}
}
\description{
Plot Lasso coefficient traces.
}
