% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso-functions.R
\name{PlotAbundaceBoxplot}
\alias{PlotAbundaceBoxplot}
\title{Plot Abundance Boxplot}
\usage{
PlotAbundaceBoxplot(abundance, coefs)
}
\arguments{
\item{abundance}{A data matrix that contains the abundance of each cluster and
corresponding outcome.}
}
\value{
A ggplot object
}
\description{
Plot abundance boxplot between outcomes.
}
