% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus-functions.R
\name{PlotClusterMFI}
\alias{PlotClusterMFI}
\title{Plot CITRUS Cluster MFI Heatmap}
\usage{
PlotClusterMFI(clustering, cluster.attributes, name, desc, ...)
}
\arguments{
\item{clustering}{A list object returned by RunCITRUS.}

\item{cluster.attributes}{A data frame returned by PlotHierarchy.}

\item{name}{Name of selected channels.}

\item{desc}{Description of selected channels.}
}
\value{
A pheatmap object.
}
\description{
Plot heatmap of mean fluroresence intensity of CITRUS clusters.
}
