% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus-functions.R
\name{PlotDist}
\alias{PlotDist}
\title{Plot Marker Distribution}
\usage{
PlotDist(clustering, name, desc, diff_cluster)
}
\arguments{
\item{clustering}{A list object from RunCITRUS}

\item{name}{Name of selected channels.}

\item{desc}{Description of selected channels.}

\item{mincluster}{Clusters}
}
\description{
Plot marker distribution of each cluster.
}
