% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso-functions.R
\name{PlotMarkerDensity}
\alias{PlotMarkerDensity}
\title{Plot Marker Density}
\usage{
PlotMarkerDensity(df, coefs)
}
\arguments{
\item{df}{A data frame with columns marker and corresponding cluster.}

\item{coefs}{The coefficient object from glmnet.}
}
\value{
A ggplot object of the marker distribution for clusters with non-zero
coefficients
}
\description{
Plot marker density plot
}
