% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenograph-functions.R
\name{PlotTSNE}
\alias{PlotTSNE}
\title{Plot TSNE using ggplot2}
\usage{
PlotTSNE(tsne_coord, cluster, sample.idx = NULL, text = NULL)
}
\arguments{
\item{tsne_coord}{A data frame of first two dimension of tsne projection.}

\item{cluster}{A character/numeric vector cluster membership corresponding to
tsne_coord}

\item{sample.idx}{A numeric/logical vector indicating selected cells}
}
\description{
Plot TSNE using ggplot2.
}
