% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PrintLaTeXTable}
\alias{PrintLaTeXTable}
\title{Print LaTex Table}
\usage{
PrintLaTeXTable(table, rownames.include = T)
}
\arguments{
\item{table}{A data matrix or data frame}

\item{rownames.include}{A logical scalar for whether to include row names}
}
\value{
A character scalar of LaTeX table
}
\description{
Print LaTeX of given table.
}
