% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-qPCR.R
\name{ProcessqPCRList}
\alias{ProcessqPCRList}
\title{Process qPCR Data Frame List}
\usage{
ProcessqPCRList(data, qPCR.idx, select.idx)
}
\arguments{
\item{data}{A list of data frame containing qPCR measurements}

\item{qPCR.idx}{A numeric vector of size indicating the index of raw
and corrected qPCR values}

\item{select.idx}{A numeric vector indicating the indices of columns to be
selected other than those in \code{qPCR.idx}}
}
\value{
A list of data frame with corrected qPCR values and selected columns
}
\description{
Process qPCR data frame in the list into the same format.
}
