% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-qPCR.R
\name{ReadExcelSheet}
\alias{ReadExcelSheet}
\title{Read Excel Sheets into a List}
\usage{
ReadExcelSheet(path, sheets, colnames = F)
}
\arguments{
\item{path}{A character scalar of full path name of the excel sheet.}

\item{sheets}{A character vector of sheet names.}

\item{colnames}{A logical scalar to indicate if first row contain column names}
}
\value{
A list of data frame of corresponding sheets
}
\description{
Read specified sheets of Excel files into a list.
}
