% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus-functions.R
\name{ReadFCS}
\alias{ReadFCS}
\title{Read an FCS file}
\usage{
ReadFCS(filePath, ...)
}
\arguments{
\item{filePath}{The full path to the FCS files.}

\item{...}{Additional arguments to flowCore::read.FCS.}
}
\value{
A flowCore::flowFrame object.
}
\description{
Read an FCS file while suppressing warnings.
}
