% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReloadPkg}
\alias{ReloadPkg}
\title{Reload Loaded Package}
\usage{
ReloadPkg(pkg)
}
\arguments{
\item{pkg}{A character scalar of package name}
}
\value{
None
}
\description{
Reload a package already loaded
}
