% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus-functions.R
\name{runCITRUS}
\alias{runCITRUS}
\title{Run CITRUS}
\usage{
runCITRUS(dataDir, selected.desc, fileList = NULL,
  family = "classification", transformCofactor = 5,
  minClusterSizePercent = 0.05, fileSampleSize = 1000, nFolds = 1,
  featureType = c("abundances"), seed = 1, pattern = "fcs",
  labels = NULL)
}
\arguments{
\item{dataDir}{Full path to the folder containing FCS files.}

\item{selected.desc}{A character vector of selected channel description.}

\item{fileList}{A list of full path to FCS files.}

\item{family}{A character scalar of model type.}

\item{transformCofactor}{An integer for arcsinh transform channel; default 5}

\item{fileSampleSize}{The number of cells subsampled from each file.}

\item{nFolds}{Number of folds used for clustering}

\item{featureType}{A character vector of feature type (abundances, proportion).}

\item{seed}{A seed to reproduce the run.}

\item{pattern}{Regular expression used to select FCS files in the folder.}

\item{labels}{Optional vector of label used for balancing clustering folds}

\item{minimumClusterSizePercent}{Minimal size of cluster in proportion.}
}
\value{
A list consisting of feature matrix, citrus clustering object and
combinedFCSSet object.
}
\description{
Run CITRUS on selected channels
}
