% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso-functions.R
\name{RunLasso}
\alias{RunLasso}
\title{Run Lasso Regression}
\usage{
RunLasso(X, y, seed = 1, ...)
}
\arguments{
\item{X}{A data matrix consists of the predictors.}

\item{y}{A numeric/factor vector of response.}
}
\value{
A list of cross validation and model object
}
\description{
Run Lasso Regression on abundance matrix provided.
}
