% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenograph-functions.R
\name{RunPhenoGraph}
\alias{RunPhenoGraph}
\title{Run PhenoGraph on a FCS File}
\usage{
RunPhenoGraph(path, desc.selected = NULL, k = 30, TSNE = T,
  seed = 1)
}
\arguments{
\item{path}{Full path to the FCS file.}

\item{desc.selected}{An optional character vector to specify description of
column selected.}

\item{k}{Nearest-neighbor parameter of PhenoGraph.}

\item{seed}{An optional seed for reproducibility.}
}
\value{
A Rphenograph object.
}
\description{
Read FCS file and select columns to run PhenoGraph on.
}
