% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus-functions.R
\name{runCITRUS}
\alias{runCITRUS}
\title{Run CITRUS}
\usage{
runCITRUS(dataDir, selected.desc, fileList = NULL,
  family = "classification", minimumClusterSizePercent = 0.05,
  fileSampleSize = 1000, nFolds = 1, featureType = c("abundances"),
  seed = 1, pattern = "fcs")
}
\arguments{
\item{dataDir}{Full path to the folder containing FCS files.}

\item{selected.desc}{A character vector of selected channel description.}

\item{fileList}{A list of full path to FCS files.}

\item{family}{A character scalar of model type.}

\item{minimumClusterSizePercent}{Minimal size of cluster in proportion.}

\item{fileSampleSize}{The number of cells subsampled from each file.}

\item{nFolds}{Number of folds used for clustering}

\item{featureType}{A character vector of feature type (abundances, proportion).}

\item{seed}{A seed to reproduce the run.}

\item{pattern}{Regular expression used to select FCS files in the folder.}
}
\value{
A list consisting of feature matrix, citrus clustering object and
combinedFCSSet object.
}
\description{
Run CITRUS on selected channels
}
